Refactor prober to use shared monitoring library
