+++ monitoring/monitorlib/requirements.txt
@@ -1,0 +1,4 @@
+google-auth==1.6.3
+requests==2.22.0
+cryptography==2.9.2
+jwcrypto==0.7
