+++ monitoring/prober/rid/test_isa_expiry.py
@@ -3,7 +3,7 @@
 import datetime
 import time
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from . import common
 from .common import SCOPE_READ, SCOPE_WRITE
