+++ monitoring/prober/rid/test_subscription_validation.py
@@ -9,7 +9,7 @@
 import datetime
 import uuid
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from . import common
 from .common import SCOPE_READ
