+++ monitoring/prober/scd/test_constraint_simple.py
@@ -10,7 +10,7 @@
 
 import datetime
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from . import common
 from .common import SCOPE_SC, SCOPE_CI, SCOPE_CM
