+++ monitoring/prober/scd/test_operation_references_error_cases.py
@@ -4,7 +4,7 @@
 import json
 import uuid
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from .common import SCOPE_SC
 
 OP_ID = '00000028-728d-40c4-8eb2-20d19c000000'
