+++ monitoring/prober/scd/test_operation_references_state_transition.py
@@ -3,7 +3,7 @@
 
 import json
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from .common import SCOPE_SC
 
 OP_ID = '00000067-cb83-4880-a7e7-1fee85000000'
