+++ monitoring/prober/scd/test_subscription_queries.py
@@ -6,7 +6,7 @@
 
 import datetime
 
-from ..infrastructure import default_scope
+from monitoring.monitorlib.infrastructure import default_scope
 from . import common
 from .common import SCOPE_SC
